In today's testing, I created four methods First Method TestLose: Here i created a Book object and then a loan object then i borrowed a Book. I then called method lose on the object. when i returned the book state to the Junit Test class, the expected result was equal to the returned result. which means our program works well.The book  must be borrowed before you can method loose on the object else it throws an illegalstateexception
Method TestDispose:Here, I created a book object and then called the
dispose method on the book. If the book were on loan, it would throw an
exception  nbut because the bookstate was available, when we returned
the state of the book, it was disposed. expected reult=returned result.

Method TestDamage: This method was a little bit more complicated because
we have to create a book instance and borrow the book, and return the
book. while returning the book we pass a boolean tru, meaning the book
has damaged. When we returned the bookstate, it showed damaged.

Method TestRepair: This method we created a book instance,borrowed the
book, and return the book. while returning the book we pass a boolean
tru, meaning the book has damaged.After that we decided to repair the
book since the state has changed to DAMAGED. If the state wasnt DAMGED,
we would have had a runtime Exception. When we returned the bookstate,
it showed AVAILABLE, because the book had been repaired.